accounts/serializers.py
from django.contrib.auth.models import Group
from rest_framework import serializers
from django.contrib.auth import get_user_model
from django.utils.translation import gettext_lazy as _
from django.contrib.auth import authenticate
from .models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = get_user_model()
        fields = ('email', 'password', 'first_name', 'last_name', 'phone', 'is_accepted')
        extra_kwargs = {
            'password': {'write_only': True}
        }

    def create(self, validated_data):
        password = validated_data.pop('password', None)
        instance = self.Meta.model(**validated_data)
        if password is not None:
            instance.set_password(password)
        instance.save()
        return instance


class AuthenticateSerializer(serializers.Serializer):
  email = serializers.CharField()
  password = serializers.CharField()

  def validate(self, data):
    # email1 = serializer.validated_data.get('email')
    if User.objects.filter(email=data['email']).exists():
        user = authenticate(**data)
        if user and user.is_active:
            return user
        raise serializers.ValidationError("Incorrect Credentials")
    raise serializers.ValidationError("Email does not exist")


class VerifySerializer(serializers.Serializer):
  email = serializers.EmailField()
  code = serializers.CharField(min_length=6, max_length=6)